requirements.txt
absl-py==2.3.1
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
fonttools==4.58.4
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.1
huggingface-hub==0.24.7
idna==3.7
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
roboflow==1.2.0
s3transfer==0.13.0
sahi==0.11.29
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.161
ultralytics-thop==2.0.14
urllib3==2.5.0
Werkzeug==3.1.3
yolov5==7.0.14
yolov8==0.0.2
